Move gatling script template into executor sources
It isn't a valid Scala file anyway, so there's no point to have it. And
it's also small enough to be included into executor's source.
